backend/requirements.txt
accelerate==1.12.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
blis==1.3.3
build==1.4.0
catalogue==2.0.10
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==1.5.2
click==8.3.1
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
cymem==2.0.13
dataclasses-json==0.6.7
Deprecated==1.3.1
distro==1.9.0
docx2txt==0.9
durationpy==0.10
ecdsa==0.19.1
emoji==2.15.0
# en_core_web_sm : installer SÉPARÉMENT avec : python -m spacy download en_core_web_sm
fastapi==0.135.1
filelock==3.25.0
filetype==1.2.0
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.2.0
google-api-core==2.30.0
google-auth==2.48.0
google-cloud-vision==3.12.1
googleapis-common-protos==1.72.0
greenlet==3.3.2
grpcio==1.78.0
grpcio-status==1.78.0
h11==0.16.0
hf-xet==1.3.2
html5lib==1.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.5.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
installer==0.7.0
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
kubernetes==35.0.0
langchain==0.3.27
langchain-chroma==0.2.6
langchain-community==0.3.31
langchain-core==0.3.83
langchain-huggingface==0.3.1
langchain-text-splitters==0.3.11
langdetect==1.0.9
langsmith==0.7.9
llvmlite==0.46.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.1
murmurhash==1.0.15
mypy_extensions==1.1.0
networkx==3.6.1
numba==0.64.0
numpy==2.4.2
oauthlib==3.3.1
olefile==0.47
onnx==1.20.1
onnxruntime==1.24.2
# onnxruntime-gpu==1.24.2  # Décommenter si GPU NVIDIA + CUDA disponible
opencv-python==4.13.0.92
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.7
overrides==7.7.0
packaging==25.0
pandas==3.0.1
pdf2image==1.17.0
pdfminer.six==20260107
pi_heif==1.3.0
pikepdf==10.3.0
pillow==12.1.1
posthog==5.4.0
preshed==3.0.12
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
psutil==7.2.2
pyasn1==0.6.2
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
pypdf==6.7.4
pypdfium2==5.5.0
PyPika==0.51.1
pyproject_hooks==1.2.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-iso639==2026.1.31
python-jose==3.5.0
python-magic==0.4.27
python-multipart==0.0.22
python-oxmsg==0.0.2
PyYAML==6.0.3
RapidFuzz==3.14.3
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.3.3
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.2.3
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
smart_open==7.5.1
soupsieve==2.8.3
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.47
srsly==2.5.2
starlette==0.52.1
sympy==1.13.1
tenacity==9.1.4
thinc==8.3.10
threadpoolctl==3.6.0
timm==1.0.25
tokenizers==0.22.2
# torch : installer manuellement selon votre configuration GPU/CPU
# CPU uniquement : pip install torch==2.6.0 torchaudio==2.6.0 torchvision==0.21.0
# GPU CUDA 12.4 : pip install torch==2.6.0+cu124 torchaudio==2.6.0+cu124 torchvision==0.21.0+cu124 --index-url https://download.pytorch.org/whl/cu124
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.3
transformers==5.2.0
typer==0.24.1
typer-slim==0.24.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
unstructured==0.21.5
unstructured-client==0.42.10
unstructured.pytesseract==0.3.15
unstructured_inference==1.2.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.41.0
wasabi==1.1.3
watchfiles==1.1.1
weasel==0.4.3
webencodings==0.5.1
websocket-client==1.9.0
websockets==16.0
wrapt==1.17.3
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
zstandard==0.25.0

streamlit>=1.40.0
